refactor: change /error-500 page to /500
