+++ app/sentry.properties
@@ -1,4 +1,3 @@
 defaults.url=https://sentry.io/
 defaults.org=bcgov-ccbc
 defaults.project=ccbc
-cli.executable=../../../.npm/_npx/a8388072043b4cbc/node_modules/@sentry/cli/bin/sentry-cli
